refactor: 피드백 반영
중복 코드 분리
