+++ study/src/test/java/thread/stage0/SynchronizationTest.java
@@ -41,7 +41,7 @@ class SynchronizationTest {
 
         private int sum = 0;
 
-        public void calculate() {
+        public synchronized void calculate() {
             setSum(getSum() + 1);
         }
