+++ tomcat/src/main/java/org/apache/coyote/http11/response/ResponseHeaders.java
@@ -1,27 +1,22 @@
 package org.apache.coyote.http11.response;
 
-import java.util.LinkedHashMap;
-import java.util.Map;
+import org.apache.coyote.http11.header.Headers;
+import org.apache.coyote.http11.header.HttpHeader;
+
 import java.util.stream.Collectors;
 
-public class ResponseHeaders {
+public class ResponseHeaders extends Headers {
 
-    private final Map<String, String> headers = new LinkedHashMap<>();
-
-    public boolean has(HttpHeader httpHeader) {
-        return headers.containsKey(httpHeader.getName());
+    public ResponseHeaders() {
+        super();
     }
 
     public void put(HttpHeader httpHeader, String header) {
-        headers.put(httpHeader.getName(), header);
-    }
-
-    public String get(HttpHeader httpHeader) {
-        return headers.get(httpHeader.getName());
+        getHeaders().put(httpHeader.getName(), header);
     }
 
     public String getMessage() {
-        return headers.entrySet().stream()
+        return getHeaders().entrySet().stream()
                 .map(entry -> entry.getKey() + ": " + entry.getValue() + " ")
                 .collect(Collectors.joining("\r\n"));
     }
